requirements.txt
# requirements.txt

aiofiles==23.1.0
aiogram==3.1.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
idna==3.4
line-profiler==4.1.2
magic-filter==1.0.12
multidict==6.0.4
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
urllib3==2.1.0
uvicorn==0.24.0.post1
yarl==1.9.2
typing-extensions~=4.7.1
redis
celerybeat-redis
